find-available-port for guile
Somehow emacs-lsp crashes right after installing lsp-guile. The reason seems
to be, that it tries to start the LSP server twice, using the same error
port number. I didn't find out the reason for this. An workaround is implemented
in this ticket: we leave the task of finding a free port to lsp-guile-connect.